small refactoring for database functions
