+++ scripts/last_geo.py
@@ -12,19 +12,6 @@ https://t.me/worth_wash_car_bot
 import psycopg2
 
 
-def connect_to_db(dbname, user, password, host) -> tuple:
-    """
-    Функция для подключения к БД
-    """
-    try:
-        conn = psycopg2.connect(dbname=dbname, user=user, password=password, host=host)
-        cur = conn.cursor()
-        return conn, cur
-    except psycopg2.Error as error:
-        print(f"Error connecting to the database: {error}")
-        return None, None
-
-
 def check_last_geo(cur, user_id) -> bool | None:
     """
     Функция принимает курсор и user_id
@@ -96,14 +83,3 @@ def update_last_geo(conn, cur, user_id, new_lat, new_lon) -> None:
     except psycopg2.Error as error:
         print(f"Error updating last geo information: {error}")
         conn.rollback()
-
-
-def close_connection_db(conn, cur) -> None:
-    """
-    Функция закрывает подключение к БД
-    """
-    try:
-        cur.close()
-        conn.close()
-    except Exception as error:
-        print(f"Error closing connection: {error}")
